requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiogram==3.15.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
backports.functools-lru-cache==1.3
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
bleach==6.2.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi==1.17.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
coloredlogs==15.0.1
comm @ file:///croot/comm_1709322850197/work
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
DAWG-Python==0.7.2
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
diffusers==0.31.0
dill==0.3.8
docopt==0.6.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.115.0
fastjsonschema==2.21.0
ffmpy==0.4.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
gradio==5.0.2
gradio_client==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
intervaltree==3.1.0
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipymarkup==0.9.0
ipython @ file:///croot/ipython_1718287989724/work
ipywidgets==8.1.5
isoduration==20.11.0
jedi @ file:///croot/jedi_1721058342488/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jiwer==3.0.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
natasha==1.6.0
navec==0.10.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx @ file:///croot/networkx_1720002482208/work
noisereduce==3.0.3
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=5f96f274d410a1682519282ae769c877d32fdbf171aa8badec7bf5e1d3a1748a
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
orjson==3.10.7
overrides==7.7.0
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.2
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peft==0.13.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///croot/pillow_1721059439630/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
pooch==1.8.2
prometheus_client==0.21.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==5.28.2
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments @ file:///croot/pygments_1684279966437/work
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==3.1.4
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-json-logger==2.0.7
python-multipart==0.0.12
pytube==15.0.0
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
RapidFuzz==3.10.0
razdel==0.5.0
referencing==0.35.1
regex==2024.7.24
requests @ file:///croot/requests_1721410876868/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.21.0
ruff==0.6.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
slovnet==0.6.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
srt==3.5.3
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.38.6
sympy==1.13.2
terminado==0.18.1
thefuzz==0.22.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.4.0
torchao==0.6.1
torchaudio==2.4.0
torchvision==0.19.0+cu124
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.5
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.46.3
triton==3.0.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.31.1
vosk==0.3.45
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.13
words2numsrus==0.1.0
xxhash==3.5.0
yargy==0.16.0
yarl==1.9.4
zipp==3.21.0
